delete node with a given value using the `.deleteItem(value)` method (#9)
* write tests in index.mjs

* add comment for clarity

* recursively call deleteFromBST helper function in deleteItem method. successfully delete leaf nodes

* successfully delete node with two children (324 node) but there is still the issue of deleting 8 (root node of the whole tree) and 4

* fix: update incorrect property used. add console.log statement for testing/debugging. successfully copy over next largest node.

however, that next largest node has not itself been deleted

* fix: delete the next largest value

i was calling the incorrect value. we should be recursively calling the next largest value (.data) at the end, in the scenario where the value for deletion has two children. after we replace the deleted node with the next largest, we delete the next largest

* remove all console log statements used in testing/debugging except for the statement that advises what the value being deleted is for clarity when running