implement iteratively the finding of the depth of a given node. fixes #13.
this makes more sense to me than the recursive version. we are traversing similar to a linked list by comparing the .data (value) of the node to be looked for against the node's .data being pointed at, and moving left or right accordingly. we also handle teh situation where the node is not found